% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggrStats.R
\name{aggrStats}
\alias{aggrStats}
\title{Aggregate statistics}
\usage{
aggrStats(in.dt, in.col, in.bycol = NULL, in.cilevel = NULL, in.rob = F)
}
\arguments{
\item{in.dt}{Input data table. Should contain at least one numeric column.}

\item{in.col}{String with the column name in the data table from which the stats should be calculated.}

\item{in.bycol}{String with the column name for grouping (optional).}

\item{in.cilevel}{Number between 0 an 1 with confidence level for t-test (optional), If not provided CIs won't be calculated.}

\item{in.rob}{Whether robust stats should be calculated instead}
}
\value{
Data Table with aggregaed statistics.
}
\description{
Calculates the number of records, mean, standard deviation, confidence interval for the mean from t-test (optionally)
from a chosen column of a data table.
}
\examples{
dt = data.table(x = c(rnorm(1000, 0, 1), rnorm(800, 1, 2)), gr = c(rep('a', 1000), rep('b', 800)))

# aggregation of the entire table without CI
dt.aggr.all = aggrStats(dt, 'x')

# aggregation within groups with CI
dt.aggr.gr = aggrStats(dt, 'x', 'gr', 0.95)
}
