% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPlateQuality.R
\name{calcPlateQuality}
\alias{calcPlateQuality}
\title{Calculate plate quality}
\usage{
calcPlateQuality(in.dt, in.col.meas, in.col.type, in.well.type,
  in.screen = c("activation", "inhibition"))
}
\arguments{
\item{in.dt}{Input data table. Rows correspond to well measurements.
Has to consist of at least 2 columns: value of the measurement in a well,
description of well types (e.g.\code{'CTRL neg'}, \code{'CTRL pos'}, \code{'Compounds'}.)}

\item{in.col.meas}{String with column name of the measurement}

\item{in.col.type}{String with column name of the well type}

\item{in.well.type}{Named list with strings that describe wells
with negaitve and positive controls, and compounds. The list has to have the following
names of elements: \code{list(ctrl.neg = 'CTRL neg', ctrl.pos = 'CTRL pos', compounds = 'Compounds')}}

\item{in.screen}{String with either of \code{c('activation', 'inhibition')}}
}
\value{
List with two items, \code{zfactor} and \code{zprime}
}
\description{
Calculates z-factor and z-prime of a measurment. Requires the actual measurment,
negative and positive controls.
}
\examples{
require(data.table)
l.well = list(compounds = 'Compounds', ctrl.neg = 'CTRL neg', ctrl.pos = 'CTRL pos')
dt = data.table(meanInt = c(rnorm(80, 2, 1),
                            rnorm(10, 0, 1),
                            rnorm(10, 2, 1)),
                wellType = c(rep(l.well$compounds, 80),
                             rep(l.well$ctrl.neg, 10),
                             rep(l.well$ctrl.pos, 10)))
calcPlateQuality(dt, 'meanInt', 'wellType', l.well, 'act')

}
