% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPlateQuality2.R
\name{calcPlateQuality2}
\alias{calcPlateQuality2}
\title{Calculate plate quality}
\usage{
calcPlateQuality2(in.dt, in.col.meas, in.col.well, in.wells.neg, in.wells.pos,
  in.wells.excl = NULL, in.screen = c("activation", "inhibition"))
}
\arguments{
\item{in.dt}{Input data table. Rows correspond to well measurements.
Has to consist of at least 2 columns: value of the measurement in a well, and well names.}

\item{in.col.meas}{Name of the column holding the measurement}

\item{in.col.well}{Name of the column holding well name}

\item{in.wells.neg}{String vector with well names of the negative control}

\item{in.wells.pos}{String vector with well names of the positive control}

\item{in.wells.excl}{String vector with additional (to neg and pos controls) well names to exclude from calculation of plate stats (mean, sd, etc); for example, untreated wells would typically be excluded. Default NULL}

\item{in.screen}{String with either of \code{c('activation', 'inhibition')}}
}
\value{
List with two items, plate and control stats, \code{zfactor}, and \code{zprime}
}
\description{
Calculates z-factor and z-prime of a measurment. Requires the actual measurment,
negative and positive controls.
}
\examples{
require(data.table)

v.wells = as.vector(outer(LETTERS[1:8],1:12,paste0))
dt = data.table(meanInt = c(rpois(76, 20), rnorm(10, 1), rnorm(10, 22)), well = v.wells)
calcPlateQuality2(dt, 'meanInt', 'well', v.wells[77:86], v.wells[87:96], 'act')

}
