% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPlateStats.R
\name{calcPlateStats}
\alias{calcPlateStats}
\title{Calculate plate statistics}
\usage{
calcPlateStats(in.dt, in.col.meas, in.col.well, in.wells.neg, in.wells.pos,
  in.wells.untr = NULL)
}
\arguments{
\item{in.dt}{Input data table with aggregate, per well values. Should contain:
numeric column with measurmeents, string column with well names (e.g. 'A01', 'A02'),
string vector with names of negative control wells, string vector with names of positive control wells.}

\item{in.col.meas}{Numeric column with measurement.}

\item{in.col.well}{Names of the column with well names (e.g. 'A01', 'A02').}

\item{in.wells.neg}{Names of wells with negative controls (e.g. 'A01', 'A02').}

\item{in.wells.pos}{Names of wells with positive controls (e.g. 'A01', 'A02').}

\item{in.wells.untr}{Names of untreated wells (e.g. 'A01', 'A02'). Default, NULL. If not NULL, untreated wells will be excluded from mean and sd for z-score calculation.}
}
\value{
Data table with same columns as original but with columns \code{npi} and \code{zscore} added.
}
\description{
Calculates NPI (normalized percentage inhibition) and z-score for a plate.
}
\examples{
require(data.table)
dt = data.table(well = c('A01', 'A02', 'A03', 'B01', 'B02', 'B03', 'C01', 'C02', 'C03'),
                x = rnorm(9))
calcPlateStats(dt, 'x', 'well', 'B01', 'C03')
}
