% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/placeLevel1st.R
\name{placeLevel1st}
\alias{placeLevel1st}
\title{Reorder levels: place a chosen factor level as first}
\usage{
placeLevel1st(in.dt, in.col, in.s)
}
\arguments{
\item{in.dt}{Input data.table}

\item{in.col}{String with the column name that contains factors to change}

\item{in.s}{String with the factor to place as first}
}
\value{
Reorders factors in a column within a dt
}
\description{
Reorder levels: place a chosen factor level as first
}
\examples{
dt = data.table(x = 1:7, gr = factor(c('a', 'a', 'b', 'b', 'b', 'c', 'c')))

placeLevel1st(dt, 'gr', 'c')
levels(dt[['gr']])
}
