% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDens.R
\name{plotDens}
\alias{plotDens}
\title{Density plot of a quantity}
\usage{
plotDens(in.dt, in.meas, in.group = NULL, in.facet = NULL,
  in.vline = NULL, in.bdepth = NULL,
  in.xlab = "\\nlog10 mean fl.int. per cell (nucleus + cytosol)")
}
\arguments{
\item{in.xlab}{}
}
\value{
ggPlot object
}
\description{
Density plot of a quantity
}
