% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDoseResp.R
\name{plotDoseResp}
\alias{plotDoseResp}
\title{Plot dose response}
\usage{
plotDoseResp(in.dt, in.xvar = "Concentration.num", in.yvar = "cytoNuc.mean",
  in.group = "Compound", in.facet = "Compound",
  in.xlab = "\\nConcentration (GDC0941 uM; Trametinib nM)",
  in.ylab = "cytoplasmic / nuclear Foxo activity\\n")
}
\arguments{
\item{in.ylab}{}
}
\value{
ggPlots2 object
}
\description{
Plot dose response
}
\examples{

}
