% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPlateHeatMap.R
\name{plotPlateHeatMap}
\alias{plotPlateHeatMap}
\title{Plot heat-map of the plate}
\usage{
plotPlateHeatMap(in.dt, in.meas, in.well, in.meas.min = NULL,
  in.meas.max = NULL, in.title.scale = "", in.wells.neg = NULL,
  in.wells.pos = NULL, in.wells.untr = NULL)
}
\arguments{
\item{in.dt}{Input data table. Should contain at least two columns: well names (e.g. 'A01', 'C03'), measurment values.}

\item{in.meas}{String of the column name with the measurement to plot.}

\item{in.well}{String of the column name with well names.}

\item{in.meas.min}{Minimum value for colour bar scale (default 'NULL'). If `NULL`, minimum is calculated from data.}

\item{in.meas.max}{Maximum value for colour bar scale (default 'NULL'). If `NULL`, maximum is calculated from data.}

\item{in.title.scale}{String with colour bar title (default '').}

\item{in.wells.neg}{String vector with names of wells with negative controls (default 'NULL'). If provided, yellow borders are drawn around those wells.}

\item{in.wells.pos}{String vector with names of wells with positive controls (default 'NULL'). If provided, green borders are drawn around those wells.}

\item{in.wells.untr}{String vector with names of untreated wells (default 'NULL'). If provided, grey borders are drawn around those wells.}
}
\value{
ggplot2 plot
}
\description{
Plots a heatmap of a chosen measured quantity. Colour scaling assigns blue gradient for negative numbers, white for zero, red gradient for positive values.
}
\examples{
require(data.table)
dt = data.table(well = c('A01', 'A02', 'A03', 'B01', 'B02', 'B03', 'C01', 'C02', 'C03'),
                npi = rnorm(9))
plotPlateHeatMap(dt, 'npi', 'well', -1.2, +1.2, 'NPI', 'B01', 'C03', 'A02')
}
