% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remOutliers.R
\name{remOutliers}
\alias{remOutliers}
\title{Remove outliers from left and/or right of the distribution}
\usage{
remOutliers(in.dt, in.colname, keeperc = 0.99, mode = c("both", "left",
  "right"))
}
\arguments{
\item{in.dt}{Input data.table}

\item{in.colname}{String with the column name to process. The column has to be of type numeric.}

\item{keeperc}{Percentage (fraction) of data to keep}

\item{mode}{Specifies where to remove data. The default \code{both} removes on both sides and keeps the inner fraction. If \code{left} (\code{right}), data on the left (right) hand side of the distribution is removed.}
}
\value{
data table without outliers. All columns are preserved.
}
\description{
Data is removed such that the inner fraction (specified by parameter \code{keeperc}) is kept
}
\examples{
require(data.table)
dt = data.table(x = rnorm(1000, 1, 0))

# keep 99\% of inner data
remOutliers(dt, 'x')

# keep 90\% of data; data will be removed from the left side of the histogram
remOutliers(dt, 'x', 0.9, 'left')
}
