% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remOutliersVec.R
\name{remOutliersVec}
\alias{remOutliersVec}
\title{Remove outliers from left and/or right of the distribution}
\usage{
remOutliersVec(in.v, keeperc = 0.99, mode = c("both", "left", "right"))
}
\arguments{
\item{keeperc}{Percentage (fraction) of data to keep}

\item{mode}{Specifies where to remove data. The default \code{both} removes on both sides and keeps the inner fraction. If \code{left} (\code{right}), data on the left (right) hand side of the distribution is removed.}

\item{in.dt}{Input vector}
}
\value{
vector without outliers. All columns are preserved.
}
\description{
Data is removed such that the inner fraction (specified by parameter \code{keeperc}) is kept
}
\examples{
v = rnorm(1000, 1, 0)

# keep 99\% of inner data
remOutliers(v)

# keep 90\% of data; data will be removed from the left side of the histogram
remOutliers(v, 0.9, 'left')
}
